+++ src/app/cli/src/app.rs
@@ -887,8 +887,13 @@ pub fn register_config_in_catalog(
     } = config.search.clone().unwrap();
 
     catalog_builder.add::<kamu_search_services::SearchServiceLocalImplLazyInit>();
+
+    let indexer = indexer.unwrap_or_default();
     catalog_builder.add_value(kamu_search_services::SearchServiceLocalIndexerConfig {
-        clear_on_start: indexer.unwrap_or_default().clear_on_start,
+        clear_on_start: indexer.clear_on_start,
+        skip_datasets_with_no_description: indexer.skip_datasets_with_no_description,
+        skip_datasets_with_no_data: indexer.skip_datasets_with_no_data,
+        payload_include_content: indexer.payload_include_content,
     });
 
     match embeddings_chunker.unwrap_or_default() {