+++ src/infra/search/qdrant/src/vector_repo_qdrant.rs
@@ -135,7 +135,7 @@ impl VectorRepository for VectorRepositoryQdrant {
 
     async fn search_points(
         &self,
-        vec: Vec<f32>,
+        vec: &[f32],
         opts: SearchPointsOpts,
     ) -> Result<Vec<FoundPoint>, InternalError> {
         use ::qdrant_client::qdrant::*;