+++ src/adapter/http/src/data/query_types.rs
@@ -13,7 +13,7 @@ use http_common::comma_separated::CommaSeparatedSet;
 use http_common::{ApiError, IntoApiError};
 use internal_error::*;
 use kamu::domain;
-use kamu_core::{DataFusionError, QueryError};
+use kamu_core::QueryError;
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -544,22 +544,9 @@ pub(crate) fn map_query_error(err: QueryError) -> ApiError {
         QueryError::DatasetNotFound(_)
         | QueryError::DatasetBlockNotFound(_)
         | QueryError::DatasetSchemaNotAvailable(_) => ApiError::not_found(err),
-        QueryError::DataFusionError(DataFusionError { source, .. }) => map_datafusion_error(source),
+        QueryError::BadQuery(_) => ApiError::bad_request(err),
         QueryError::Access(err) => err.api_err(),
         QueryError::Internal(err) => err.api_err(),
-    }
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-fn map_datafusion_error(err: datafusion::error::DataFusionError) -> ApiError {
-    match err {
-        datafusion::error::DataFusionError::SQL(err, _) => ApiError::bad_request(err),
-        datafusion::error::DataFusionError::Plan(_) => ApiError::bad_request(err),
-        datafusion::error::DataFusionError::Diagnostic(_diag, inner) => {
-            map_datafusion_error(*inner)
-        }
-        _ => err.int_err().api_err(),
     }
 }