+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Recommendation: for ease of reading, use the following order:
 ### Added
 - More configuration options for semantic search indexing, allowing to skip datasets without descriptions or no data.
 ### Changed
+- Semantic search service will make repeated queries to the vector store to return the desired number of datasets per results page.
 - Flow: Updated the `BatchingRule` trigger to accept 0 for both properties(`min_records_to_await` and `max_batching_interval`), enabling dependency flow execution even when no data is added to the root dataset.
 
 ## [0.228.0] - 2025-03-19